DigitalRainScreen: Refactor to separate digital rain simulation logic from screen-specific UI, layout, and diagnostics.
Modifies:
- TUI.vcxproj
- TUI.vcxproj.filters
- Screens/DigitalRainScreen.h/.cpp

Adds:
- Screens/Content/DigitalRainEffect.h/.cpp

Key changes:

- Introduced DigitalRainEffect (Screens/Content/DigitalRainEffect.h/.cpp)
  - Encapsulates all rain simulation and rendering logic:
    - stream state and lifecycle
    - glyph pool selection based on TerminalHostKind
    - dead glyph spawning and management
    - RNG-driven animation and mutation
    - trail styling and composition rules
  - Renders into a caller-provided viewport (Rect)

- Simplified DigitalRainScreen (Screens/DigitalRainScreen.h/.cpp)
  - Owns all UI/chrome:
    - frame and panel layout (PageComposer regions)
    - title and footer rendering
    - glyph preview strip
    - Unicode validation panel
    - DigiXRay overlay
  - Defines viewport region and passes it to DigitalRainEffect::draw()

- Removed duplicated state and logic
  - Eliminated glyph pool duplication in DigitalRainScreen
  - Screen now queries effect for glyph pool when needed (preview UI)
  - Dead glyph state is now fully owned by DigitalRainEffect

- Enables reuse of DigitalRainEffect in:
  - windows
  - popups
  - split-screen layouts
  - future widget system
