+++ TUI/Screens/DigitalRainScreen.cpp
@@ -7,6 +7,7 @@
 
 #include "Core/Rect.h"
 #include "Rendering/Composition/WritePresets.h"
+#include "Rendering/Composition/PageComposer.h"
 #include "Rendering/Objects/TextObjectComposer.h"
 #include "Rendering/Objects/TextObjectFactory.h"
 #include "Rendering/Objects/TextObjectExporter.h"
@@ -25,27 +26,14 @@
 
 namespace
 {
-    constexpr char32_t RabbitGlyph = U'\U0001F407';
+    using Composition::Alignment;
+    using Composition::PageComposer;
 
     constexpr int MinimumScreenWidth = 48;
     constexpr int MinimumScreenHeight = 14;
     constexpr int FooterRows = 2;
     constexpr double PreviewAdvanceIntervalSeconds = 0.08;
-    constexpr double MutationIntervalSeconds = 0.065;
-
-    constexpr int MaximumDigitalDrops = 150;
-    constexpr int DeadGlyphsSpawnRate = 25;
-    constexpr int DeadGlyphBlinkRate = 10;
-
-    struct DeadGlyph
-    {
-        int x = 0;
-        int y = 0;
-        char32_t glyph = U' ';
-    };
-
-    std::vector<DeadGlyph> m_deadGlyphs;
-    double m_deadGlyphSpawnAccumulator = 0.0;
+    constexpr char32_t RabbitGlyph = U'\U0001F407';
 }
 
 rainXRay::rainXRay()
@@ -98,13 +86,23 @@ void rainXRay::draw(ScreenBuffer& buffer)
 
 void rainXRay::move(int screenWidth, int screenHeight, double deltaTime)
 {
-    m_xRayX += m_xRayDx * static_cast<float>(deltaTime);
-    m_xRayY += m_xRayDy * static_cast<float>(deltaTime);
+    if (!m_xRay_initialized || screenWidth <= 0 || screenHeight <= 0)
+    {
+        return;
+    }
 
     const float minX = static_cast<float>(m_LeftBounds);
     const float maxX = static_cast<float>(screenWidth - m_RightBounds - m_xRayWidth);
     const float minY = static_cast<float>(m_TopBounds);
     const float maxY = static_cast<float>(screenHeight - m_BottomBounds - m_xRayHeight);
+
+    if (maxX <= minX || maxY <= minY)
+    {
+        return;
+    }
+
+    m_xRayX += m_xRayDx * static_cast<float>(deltaTime);
+    m_xRayY += m_xRayDy * static_cast<float>(deltaTime);
 
     if (m_xRayX <= minX || m_xRayX >= maxX)
     {
@@ -159,85 +157,11 @@ bool rainXRay::initialized()
     return m_xRay_initialized;
 }
 
-std::u32string DigitalRainScreen::buildConsoleGlyphPool()
-{
-    return std::u32string(
-        U"0123456789A₿CDEFGHIJKLMNOPQRSTUVWXYZ"
-        U"@#$%&*+=<>"
-        U"[]{}()"
-        U"█▓▒░"
-        U"■□▪▫"
-        U"▲▼"
-        U"○●"
-        U"ΑβϲδεφϑհιյΚλʍƞɸπθʀστυƔѡϰψȥ"
-        U"┌┐└┘├┤┬┴┼"
-        U"╔╗╚╝╠╣╦╩╬"
-        U"│┃─━"
-        U"←↑→↓");
-}
-
-//        U"♜♞♝♛♚♝♞♜♖♘♗♕♔♗♘♖"
-//        + std::u32string(1, RabbitGlyph)
-std::u32string DigitalRainScreen::buildTerminalGlyphPool()
-{
-    return std::u32string(
-        U"アァカサタナハマヤャラワガザダバパ"
-        U"イィキシチニヒミリヰギジヂビピ"
-        U"ウゥクスツヌフムユュルグズブヅプ"
-        U"エェケセテネヘメレヱゲゼデベペ"
-        U"オォコソトノホモヨョロヲゴゾドボポヴッン"
-        U"0123456789ABCDEFGHIJKLMNOPQRSTUVWXYZ" // standard numbers and letters
-        U"ΑβϲδεφϑհιյΚλʍƞɸπθʀστυƔѡϰψȥ"           // greek alphabet
-        U"♪♫⌘₿äü∄∃ƒ±£µℇ");
-}
-
-std::u32string DigitalRainScreen::buildGlyphPoolForHost(TerminalHostKind hostKind)
-{
-    switch (hostKind)
-    {
-    case TerminalHostKind::ClassicConsoleWindow:
-    case TerminalHostKind::RedirectedOrPipe:
-    case TerminalHostKind::Unknown:
-        return buildConsoleGlyphPool();
-
-    case TerminalHostKind::WindowsTerminal:
-    case TerminalHostKind::VsCodeIntegratedTerminal:
-    case TerminalHostKind::ConEmu:
-    case TerminalHostKind::Mintty:
-    case TerminalHostKind::OtherTerminalHost:
-    default:
-        return buildTerminalGlyphPool();
-    }
-}
-
 DigitalRainScreen::DigitalRainScreen(TerminalHostKind hostKind)
     : m_hostKind(hostKind)
-    , m_glyphPool(buildGlyphPoolForHost(hostKind))
-    , m_rng(0x4D415452u)
+    , m_DigitalRain(hostKind)
 {
     const Color black = Color::FromBasic(Color::Basic::Black);
-
-    m_headStyle =
-        style::Bold
-        + style::Fg(Color::FromBasic(Color::Basic::BrightWhite))
-        + style::Bg(black);
-
-    m_secondStyle =
-        style::Fg(Color::FromBasic(Color::Basic::White))
-        + style::Bg(black);
-
-    m_thirdStyle =
-        style::Bold
-        + style::Fg(Color::FromBasic(Color::Basic::BrightGreen))
-        + style::Bg(black);
-
-    m_fourthStyle =
-        style::Fg(Color::FromBasic(Color::Basic::Green))
-        + style::Bg(black);
-
-    m_tailStyle =
-        style::Fg(Color::FromBasic(Color::Basic::BrightBlack))
-        + style::Bg(black);
 
     m_backgroundStyle = style::Bg(black);
 
@@ -258,6 +182,12 @@ DigitalRainScreen::DigitalRainScreen(TerminalHostKind hostKind)
     m_borderStyle =
         style::Fg(Color::FromBasic(Color::Basic::Green))
         + style::Bg(black);
+
+    m_windowStyle =
+        style::Fg(Color::FromBasic(Color::Basic::BrightBlack))
+        + style::Bg(black);
+
+    m_backgroundStyle = style::Bg(black);
 }
 
 void DigitalRainScreen::onEnter()
@@ -265,26 +195,19 @@ void DigitalRainScreen::onEnter()
     m_elapsedSeconds = 0.0;
     m_previewAdvanceTimer = 0.0;
     m_previewOffset = 0;
-    m_streams.clear();
 
     invalidateStaticUiCache();
     invalidateMinimumScreenUiCache();
     m_xRayMachine.invalidateXray();
+    m_DigitalRain.onEnter();
 }
 
 void DigitalRainScreen::update(double deltaTime)
 {
     m_elapsedSeconds += deltaTime;
-
-    if (m_rainWidth <= 0 || m_rainHeight <= 0 || m_streams.empty())
-    {
-        return;
-    }
-
-    updateStreams(deltaTime);
-    spawnDeadGlyphs(deltaTime);
     updatePreview(deltaTime);
     m_xRayMachine.move(m_screenWidth, m_screenHeight, deltaTime);
+    m_DigitalRain.update(deltaTime);
 }
 
 void DigitalRainScreen::draw(Surface& surface)
@@ -302,9 +225,18 @@ void DigitalRainScreen::draw(Surface& surface)
     }
 
     ensureLayout(screenWidth, screenHeight);
+
     ensureStaticUiCache();
 
-    drawRain(buffer);
+    PageComposer page(buffer);
+    page.clearRegions();
+
+    page.createFullScreenRegion("Screen");
+    page.createInsetRegion("DigitalRain", "Screen", 3, 2, 3, 3);
+
+    const Rect& viewport = page.resolveRegion("DigitalRain");
+    m_DigitalRain.draw(surface, viewport);
+
     m_staticUiObject.draw(buffer, 0, 0, Composition::WritePresets::visibleObject());
 
     const int footerLabelY = m_screenHeight - 3;
@@ -324,29 +256,16 @@ void DigitalRainScreen::draw(Surface& surface)
 
 void DigitalRainScreen::ensureLayout(int screenWidth, int screenHeight)
 {
-    const int newRainLeft = 3;
-    const int newRainTop = 2;
-    const int newRainWidth = std::max(0, screenWidth - 7);
-    const int newRainHeight = std::max(0, screenHeight - 4 - FooterRows);
-
-    if (screenWidth == m_screenWidth &&
-        screenHeight == m_screenHeight &&
-        newRainWidth == m_rainWidth &&
-        newRainHeight == m_rainHeight)
+    if (screenWidth == m_screenWidth && screenHeight == m_screenHeight)
     {
         return;
     }
 
     m_screenWidth = screenWidth;
     m_screenHeight = screenHeight;
-    m_rainLeft = newRainLeft;
-    m_rainTop = newRainTop;
-    m_rainWidth = newRainWidth;
-    m_rainHeight = newRainHeight;
 
     invalidateStaticUiCache();
     m_xRayMachine.invalidateXray();
-    rebuildStreams();
 }
 
 void DigitalRainScreen::ensureStaticUiCache()
@@ -420,7 +339,7 @@ TextObject DigitalRainScreen::buildStaticUiTextObject() const
         1,
         m_screenWidth - 4,
         m_screenHeight - 2 - FooterRows,
-        m_tailStyle,
+        m_windowStyle,
         ObjectFactory::singleLineBorder(),
         10,
         "innerFrame");
@@ -430,7 +349,7 @@ TextObject DigitalRainScreen::buildStaticUiTextObject() const
     const int footerPreviewY = m_screenHeight - 2;
 
     composer.addText(U"Pool", 2, footerLabelY, m_labelStyle, 20, "poolLabel");
-    composer.addGlyph(U':', 6, footerLabelY, m_headStyle, 20, "poolColon");
+    composer.addGlyph(U':', 6, footerLabelY, m_labelStyle, 20, "poolColon");
 
     if (usesConsoleFooter())
     {
@@ -508,7 +427,7 @@ TextObject DigitalRainScreen::buildUnicodeContractPanelTextObject() const
         0,
         panelWidth,
         panelHeight,
-        m_tailStyle,
+        m_windowStyle,
         ObjectFactory::singleLineBorder(),
         10,
         "contractPanelFrame");
@@ -751,151 +670,13 @@ bool DigitalRainScreen::usesConsoleFooter() const
     }
 }
 
-void DigitalRainScreen::rebuildStreams()
-{
-    if (m_rainWidth <= 0 || m_rainHeight <= 0)
-    {
-        m_streams.clear();
-        return;
-    }
 
-    m_streams.assign(static_cast<std::size_t>(m_rainWidth), Stream{});
-
-    for (Stream& stream : m_streams)
-    {
-        resetStream(stream, true);
-    }
-}
-
-void DigitalRainScreen::resetStream(Stream& stream, bool staggerStart)
-{
-    stream.active = false;
-    stream.headY = 0.0f;
-    stream.speed = 0.0f;
-    stream.length = 0;
-    stream.mutationTimer = 0.0;
-    stream.glyphs.clear();
-    stream.composition = DropComposition::GreenOnly;
-
-    std::uniform_real_distribution<double> delayDistribution(
-        0.0,
-        staggerStart ? 4.8 : 2.5);
-
-    stream.respawnDelay = delayDistribution(m_rng);
-
-    if (!staggerStart)
-    {
-        configureActiveStream(stream);
-    }
-}
-
-void DigitalRainScreen::configureActiveStream(Stream& stream)
-{
-    std::uniform_real_distribution<float> speedDistribution(0.25f, 28.0f);
-    std::uniform_int_distribution<int> lengthDistribution(6, std::max(6, std::min(m_rainHeight, 24)));
-    std::uniform_real_distribution<float> headOffsetDistribution(1.0f, static_cast<float>(std::max(2, m_rainHeight / 2)));
-    std::uniform_int_distribution<int> compositionDistribution(0, 99);
-
-    stream.active = true;
-    stream.speed = speedDistribution(m_rng);
-    stream.length = lengthDistribution(m_rng);
-    stream.headY = -headOffsetDistribution(m_rng);
-    stream.mutationTimer = 0.0;
-    stream.glyphs.assign(static_cast<std::size_t>(stream.length), U' ');
-
-    const int compositionRoll = compositionDistribution(m_rng);
-
-    if (compositionRoll < 25)
-    {
-        stream.composition = DropComposition::BrightWhiteWhite;
-    }
-    else if (compositionRoll < 50)
-    {
-        stream.composition = DropComposition::WhiteOnly;
-    }
-    else
-    {
-        stream.composition = DropComposition::GreenOnly;
-    }
-
-    for (char32_t& glyph : stream.glyphs)
-    {
-        glyph = randomGlyph();
-    }
-}
-
-void DigitalRainScreen::updateStreams(double deltaTime)
-{
-    std::uniform_int_distribution<int> mutationCountDistribution(1, 3);
-
-    for (Stream& stream : m_streams)
-    {
-        if (!stream.active)
-        {
-            stream.respawnDelay -= deltaTime;
-
-            if (stream.respawnDelay <= 0.0)
-            {
-                // Enforce global active stream cap
-                if (countActiveStreams() < MaximumDigitalDrops)
-                {
-                    configureActiveStream(stream);
-                }
-                else
-                {
-                    // Defer activation slightly to avoid tight retry loop
-                    stream.respawnDelay = 0.1;
-                }
-            }
-
-            continue;
-        }
-
-        const int previousHeadRow = static_cast<int>(std::floor(stream.headY));
-        stream.headY += stream.speed * static_cast<float>(deltaTime);
-        const int currentHeadRow = static_cast<int>(std::floor(stream.headY));
-
-        for (int row = previousHeadRow; row < currentHeadRow; ++row)
-        {
-            if (!stream.glyphs.empty())
-            {
-                for (int i = stream.length - 1; i > 0; --i)
-                {
-                    stream.glyphs[static_cast<std::size_t>(i)] = stream.glyphs[static_cast<std::size_t>(i - 1)];
-                }
-
-                stream.glyphs[0] = randomGlyph();
-            }
-        }
-
-        stream.mutationTimer += deltaTime;
-
-        while (stream.mutationTimer >= MutationIntervalSeconds)
-        {
-            stream.mutationTimer -= MutationIntervalSeconds;
-
-            const int mutationCount = mutationCountDistribution(m_rng);
-            if (!stream.glyphs.empty())
-            {
-                std::uniform_int_distribution<int> glyphIndexDistribution(0, static_cast<int>(stream.glyphs.size()) - 1);
-
-                for (int i = 0; i < mutationCount; ++i)
-                {
-                    stream.glyphs[static_cast<std::size_t>(glyphIndexDistribution(m_rng))] = randomGlyph();
-                }
-            }
-        }
-
-        if (stream.headY - static_cast<float>(stream.length) > static_cast<float>(m_rainHeight))
-        {
-            resetStream(stream, false);
-        }
-    }
-}
 
 void DigitalRainScreen::updatePreview(double deltaTime)
 {
-    if (m_glyphPool.empty())
+    std::u32string glyphPool = m_DigitalRain.getGlyphPool();
+
+    if (glyphPool.empty())
     {
         return;
     }
@@ -905,127 +686,25 @@ void DigitalRainScreen::updatePreview(double deltaTime)
     while (m_previewAdvanceTimer >= PreviewAdvanceIntervalSeconds)
     {
         m_previewAdvanceTimer -= PreviewAdvanceIntervalSeconds;
-        m_previewOffset = (m_previewOffset + 1) % static_cast<int>(m_glyphPool.size());
-    }
-}
-
-int DigitalRainScreen::countActiveStreams() const
-{
-    int count = 0;
-
-    for (const Stream& stream : m_streams)
-    {
-        if (stream.active)
-        {
-            ++count;
-        }
-    }
-
-    return count;
-}
-
-void DigitalRainScreen::spawnDeadGlyphs(double deltaTime)
-{
-    // Accumulate time
-    m_deadGlyphSpawnAccumulator += deltaTime;
-
-    // How many glyphs per second
-    constexpr double spawnRate = static_cast<double>(DeadGlyphsSpawnRate);
-
-    if (spawnRate <= 0.0)
-    {
-        return;
-    }
-
-    const double interval = 1.0 / spawnRate;
-
-    // Spawn as many as needed this frame
-    while (m_deadGlyphSpawnAccumulator >= interval)
-    {
-        m_deadGlyphSpawnAccumulator -= interval;
-
-        if (m_rainWidth <= 0 || m_rainHeight <= 0)
-        {
-            continue;
-        }
-
-        std::uniform_int_distribution<int> xDist(0, m_rainWidth - 1);
-        std::uniform_int_distribution<int> yDist(0, m_rainHeight - 1);
-
-        DeadGlyph glyph;
-        glyph.x = xDist(m_rng);
-        glyph.y = yDist(m_rng);
-        glyph.glyph = randomGlyph();
-
-        m_deadGlyphs.push_back(glyph);
-    }
-
-    // Optional: cap to prevent unbounded growth
-    constexpr std::size_t maxDeadGlyphs = 512;
-    if (m_deadGlyphs.size() > maxDeadGlyphs)
-    {
-        m_deadGlyphs.erase(
-            m_deadGlyphs.begin(),
-            m_deadGlyphs.begin() + (m_deadGlyphs.size() - maxDeadGlyphs));
-    }
-}
-
-void DigitalRainScreen::drawRain(ScreenBuffer& buffer) const
-{
-    buffer.fillRect(
-        Rect{ Point{ m_rainLeft, m_rainTop }, Size{ m_rainWidth, m_rainHeight } },
-        U' ',
-        m_backgroundStyle);
-
-    // Draw dead glyphs (background noise layer)
-    for (const DeadGlyph& glyph : m_deadGlyphs)
-    {
-        buffer.writeCodePoint(
-            m_rainLeft + glyph.x,
-            m_rainTop + glyph.y,
-            glyph.glyph,
-            m_tailStyle); // dim / "dead"
-    }
-
-    for (int column = 0; column < m_rainWidth; ++column)
-    {
-        const Stream& stream = m_streams[static_cast<std::size_t>(column)];
-        if (!stream.active)
-        {
-            continue;
-        }
-
-        const int headRow = static_cast<int>(std::floor(stream.headY));
-
-        for (int trailIndex = 0; trailIndex < stream.length; ++trailIndex)
-        {
-            const int y = headRow - trailIndex;
-
-            if (y < 0 || y >= m_rainHeight)
-            {
-                continue;
-            }
-
-            const char32_t glyph = stream.glyphs[static_cast<std::size_t>(trailIndex)];
-            const Style style = styleForTrailIndex(trailIndex, stream.length, stream.composition);
-            buffer.writeCodePoint(m_rainLeft + column, m_rainTop + y, glyph, style);
-        }
+        m_previewOffset = (m_previewOffset + 1) % static_cast<int>(glyphPool.size());
     }
 }
 
 void DigitalRainScreen::drawPreviewLine(ScreenBuffer& buffer, int x, int y, int availableWidth, int startIndex) const
 {
-    if (availableWidth <= 0 || m_glyphPool.empty())
+    std::u32string glyphPool = m_DigitalRain.getGlyphPool();
+
+    if (availableWidth <= 0 || glyphPool.empty())
     {
         return;
     }
 
     int writeX = x;
-    int glyphIndex = startIndex % static_cast<int>(m_glyphPool.size());
+    int glyphIndex = startIndex % static_cast<int>(glyphPool.size());
 
     while (writeX < x + availableWidth)
     {
-        const char32_t glyph = m_glyphPool[static_cast<std::size_t>(glyphIndex)];
+        const char32_t glyph = glyphPool[static_cast<std::size_t>(glyphIndex)];
         const CellWidth width = UnicodeWidth::measureCodePointWidth(glyph);
         const int glyphWidth = (width == CellWidth::Two) ? 2 : 1;
 
@@ -1036,96 +715,6 @@ void DigitalRainScreen::drawPreviewLine(ScreenBuffer& buffer, int x, int y, int 
 
         buffer.writeCodePoint(writeX, y, glyph, m_previewStyle);
         writeX += glyphWidth;
-        glyphIndex = (glyphIndex + 1) % static_cast<int>(m_glyphPool.size());
+        glyphIndex = (glyphIndex + 1) % static_cast<int>(glyphPool.size());
     }
-}
-
-char32_t DigitalRainScreen::randomGlyph()
-{
-    if (m_glyphPool.empty())
-    {
-        return U' ';
-    }
-
-    std::uniform_int_distribution<int> glyphDistribution(0, static_cast<int>(m_glyphPool.size()) - 1);
-    return m_glyphPool[static_cast<std::size_t>(glyphDistribution(m_rng))];
-}
-
-Style DigitalRainScreen::styleForTrailIndex(
-    int trailIndex,
-    int streamLength,
-    DropComposition composition) const
-{
-    if (streamLength <= 0)
-    {
-        return m_tailStyle;
-    }
-
-    const int brightBlackStart = std::max(0, streamLength - 2);
-    if (trailIndex >= brightBlackStart)
-    {
-        return m_tailStyle;
-    }
-
-    int brightGreenStart = 0;
-    int greenStart = 0;
-
-    switch (composition)
-    {
-    case DropComposition::BrightWhiteWhite:
-        // 1 BrightWhite
-        // 1 White
-        // half middle BrightGreen
-        // half middle Green
-        // last 2 BrightBlack
-        if (trailIndex == 0)
-        {
-            return m_headStyle;
-        }
-
-        if (trailIndex == 1)
-        {
-            return m_secondStyle;
-        }
-
-        brightGreenStart = 2;
-        break;
-
-    case DropComposition::WhiteOnly:
-        // 1 White
-        // half middle BrightGreen
-        // half middle Green
-        // last 2 BrightBlack
-        if (trailIndex == 0)
-        {
-            return m_secondStyle;
-        }
-
-        brightGreenStart = 1;
-        break;
-
-    case DropComposition::GreenOnly:
-    default:
-        // half BrightGreen
-        // half Green
-        // last 2 BrightBlack
-        brightGreenStart = 0;
-        break;
-    }
-
-    const int middleCount = std::max(0, brightBlackStart - brightGreenStart);
-    if (middleCount <= 0)
-    {
-        return m_fourthStyle;
-    }
-
-    const int brightGreenCount = (middleCount + 1) / 2;
-    greenStart = brightGreenStart + brightGreenCount;
-
-    if (trailIndex < greenStart)
-    {
-        return m_thirdStyle;
-    }
-
-    return m_fourthStyle;
 }
